internal/httpsvc: new package
Introduce a new httpsvc package to extract the HTTP server which
respects the workgroup.Group API facility.
