+++ internal/debug/debug.go
@@ -16,67 +16,33 @@
 package debug
 
 import (
-	"context"
 	"fmt"
 	"net/http"
 	"net/http/pprof"
-	"time"
 
 	"github.com/heptio/contour/internal/dag"
+	"github.com/heptio/contour/internal/httpsvc"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/sirupsen/logrus"
 )
 
 // Service serves various http endpoints including /debug/pprof.
 type Service struct {
-	Addr string
-	Port int
-
-	logrus.FieldLogger
+	httpsvc.Service
 
 	*dag.Builder
 }
 
 // Start fulfills the g.Start contract.
 // When stop is closed the http server will shutdown.
-func (svc *Service) Start(stop <-chan struct{}, registry *prometheus.Registry) (err error) {
-	defer func() {
-		if err != nil {
-			svc.WithError(err).Error("terminated with error")
-		} else {
-			svc.Info("stopped")
-		}
-	}()
-	mux := http.NewServeMux()
-	registerProfile(mux)
-	registerHealthCheck(mux)
-	registerMetrics(mux, registry)
+func (svc *Service) Start(stop <-chan struct{}, registry *prometheus.Registry) error {
+	registerProfile(&svc.ServeMux)
+	registerHealthCheck(&svc.ServeMux)
+	registerMetrics(&svc.ServeMux, registry)
 
 	// register DAG dot writer.
-	mux.HandleFunc("/debug/dag", svc.writeDot)
-
-	s := http.Server{
-		Addr:           fmt.Sprintf("%s:%d", svc.Addr, svc.Port),
-		Handler:        mux,
-		ReadTimeout:    10 * time.Second,
-		WriteTimeout:   5 * time.Minute, // allow for long trace requests
-		MaxHeaderBytes: 1 << 11,         // 8kb should be enough for anyone
-	}
-
-	go func() {
-		// wait for stop signal from group.
-		<-stop
-
-		// shutdown the server with 5 seconds grace.
-		ctx := context.Background()
-		ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
-		defer cancel()
-		s.Shutdown(ctx)
-	}()
-
-	svc.WithField("address", s.Addr).Info("started")
-	return s.ListenAndServe()
+	svc.ServeMux.HandleFunc("/debug/dag", svc.writeDot)
+	return svc.Service.Start(stop)
 }
 
 func registerProfile(mux *http.ServeMux) {
